+++ src/ids-locale/cultures/he-IL.js
@@ -1,0 +1,64 @@
+// Info Comes from http://www.unicode.org/Public/cldr/25/
+const locale = {
+  // layout/language
+  language: 'he',
+  englishName: 'Hebrew (Israel)',
+  nativeName: 'עברית (ישראל)',
+  // layout/orientation/@characters
+  direction: 'right-to-left',
+  // ca-gregorian
+  calendars: [{
+    name: 'gregorian',
+    // ca-gregorian/main/dates/calendars/gregorian/dateFormats/
+    dateFormat: {
+      separator: '/', // Infered
+      timeSeparator: ':',
+      short: 'dd/MM/yyyy', // use four digit year
+      medium: 'd בMMM yyyy',
+      long: 'd בMMMM yyyy',
+      full: 'EEEE, d בMMMM yyyy',
+      month: 'd MMMM',
+      year: 'MMMM yyyy',
+      timestamp: 'H:mm:ss',
+      timestampMillis: 'H:mm:ss.SSS',
+      dayOfWeek: 'EEE d',
+      hour: 'H:mm',
+      datetime: 'dd/MM/yyyy H:mm',
+      datetimeMillis: 'dd/MM/yyyy H:mm:ss.SSS',
+      timezone: 'dd/MM/yyyy H:mm zz',
+      timezoneLong: 'dd/MM/yyyy H:mm zzzz'
+    }, // Infered short + short gregorian/dateTimeFormats
+    // ca-gregorian/main/dates/calendars/gregorian/days/format/short or abbreviated (2 digit)
+    days: {
+      wide: ['יום ראשון', 'יום שני', 'יום שלישי', 'יום רביעי', 'יום חמישי', 'יום שישי', 'יום שבת'],
+      abbreviated: ['יום א׳', 'יום ב׳', 'יום ג׳', 'יום ד׳', 'יום ה׳', 'יום ו׳', 'שבת'],
+      narrow: ['א׳', 'ב׳', 'ג׳', 'ד׳', 'ה׳', 'ו׳', 'ש׳']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/months/format/wide
+    months: {
+      wide: ['ינואר', 'פברואר', 'מרץ', 'אפריל', 'מאי', 'יוני', 'יולי', 'אוגוסט', 'ספטמבר', 'אוקטובר', 'נובמבר', 'דצמבר'],
+      abbreviated: ['ינו׳', 'פבר׳', 'מרץ', 'אפר׳', 'מאי', 'יוני', 'יולי', 'אוג׳', 'ספט׳', 'אוק׳', 'נוב׳', 'דצמ׳']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/timeFormats/short
+    timeFormat: 'H:mm',
+    // ca-gregorian/main/dates/calendars/gregorian/dayPeriods/wide
+    dayPeriods: ['לפנה״צ', 'אחה״צ'],
+    firstDayofWeek: 0 // Starts on Sun
+  }],
+  // numbers/currencyFormats-numberSystem-latn/standard
+  currencySign: '₪',
+  currencyFormat: '### ¤',
+  // numbers/symbols-numberSystem-latn
+  numbers: {
+    percentSign: '%',
+    percentFormat: '### %',
+    percentSuffix: ' %',
+    percentPrefix: '',
+    minusSign: '-',
+    decimal: '.',
+    group: ',',
+    groupSizes: [3, 3]
+  }
+};
+
+export { locale };