+++ src/ids-locale/cultures/uk-UA.js
@@ -1,0 +1,64 @@
+// Info Comes from http://www.unicode.org/Public/cldr/25/
+const locale = {
+  // layout/language
+  language: 'uk',
+  englishName: 'Ukrainian (Ukraine)',
+  nativeName: 'україньска (Україна)',
+  // layout/orientation/@characters
+  direction: 'left-to-right',
+  // ca-gregorian
+  calendars: [{
+    name: 'gregorian',
+    // ca-gregorian/main/dates/calendars/gregorian/dateFormats/
+    dateFormat: {
+      separator: '.', // Infered
+      timeSeparator: ':',
+      short: 'dd.MM.yyyy', // use four digit year
+      medium: 'd MMM yyyy р.',
+      long: 'd MMMM yyyy р.',
+      full: 'EEEE, d MMMM yyyy р.',
+      month: 'd MMMM',
+      year: 'MMMM yyyy  р.',
+      dayOfWeek: 'EEE d',
+      timestamp: 'HH:mm:ss',
+      timestampMillis: 'HH:mm:ss.SSS',
+      hour: 'HH:mm',
+      datetime: 'dd.MM.yyyy HH:mm',
+      datetimeMillis: 'dd.MM.yyyy HH:mm:ss.SSS',
+      timezone: 'dd.MM.yyyy HH:mm zz',
+      timezoneLong: 'dd.MM.yyyy HH:mm zzzz'
+    }, // Infered short + short gregorian/dateTimeFormats
+    // ca-gregorian/main/dates/calendars/gregorian/days/format/short or abbreviated (2 digit)
+    days: {
+      wide: ['неділя', 'понеділок', 'вівторок', 'середа', 'четвер', 'пʼятниця', 'субота'],
+      abbreviated: ['нд', 'пн', 'вт', 'ср', 'чт', 'пт', 'сб'],
+      narrow: ['H', 'П', 'B', 'С', 'Ч', 'П', 'C']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/months/format/wide
+    months: {
+      wide: ['січень', 'лютий', 'березень', 'квітень', 'травень', 'червень', 'липень', 'серпень', 'вересень', 'жовтень', 'листопад', 'грудень'],
+      abbreviated: ['січ', 'лют', 'бер', 'кві', 'тра', 'чер', 'лип', 'сер', 'вер', 'жов', 'лис', 'гру']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/timeFormats/short
+    timeFormat: 'HH:mm',
+    // ca-gregorian/main/dates/calendars/gregorian/dayPeriods/wide
+    dayPeriods: ['дп', 'пп'],
+    firstDayofWeek: 1 // Starts on Mon
+  }],
+  // numbers/currencyFormats-numberSystem-latn/standard
+  currencySign: '₴',
+  currencyFormat: '### ¤',
+  // numbers/symbols-numberSystem-latn
+  numbers: {
+    percentSign: '%',
+    percentFormat: '###%',
+    percentSuffix: '%',
+    percentPrefix: '',
+    minusSign: '-',
+    decimal: ',',
+    group: ' ',
+    groupSizes: [3, 3]
+  }
+};
+
+export { locale };