+++ src/ids-progress-bar/ids-progress-bar.js
@@ -31,8 +31,8 @@ class IdsProgressBar extends mix(IdsElement).with(IdsEventsMixin) {
   }
 
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
     return [