+++ src/ids-tabs/ids-tabs.js
@@ -37,8 +37,8 @@ export default class IdsTabs extends mix(IdsElement).with(
   }
 
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
     return [attributes.ORIENTATION, attributes.VALUE];